refactor: move ugrid.c => tui/ugrid.c
This module is only used by the tui code.
