+++ src/nvim/tui/tui.c
@@ -37,8 +37,8 @@
 #include "nvim/tui/input.h"
 #include "nvim/tui/terminfo.h"
 #include "nvim/tui/tui.h"
+#include "nvim/tui/ugrid.h"
 #include "nvim/types_defs.h"
-#include "nvim/ugrid.h"
 #include "nvim/ui_client.h"
 #include "nvim/ui_defs.h"
 #include "nvim/vim_defs.h"
